* commit '1665af1c58959fd69ff42459a1d723f5cbf360dd':
  change native time to double to allow sub-ms resolution

+++ suite/pts/deviceTests/dram/src/com/android/pts/dram/MemoryNative.java
@@ -27,5 +27,5 @@ public class MemoryNative {
      * @param repeatition
      * @return time spent in copying in ms.
      */
-    public static native long runMemcpy(int bufferSize, int repetition);
+    public static native double runMemcpy(int bufferSize, int repetition);
 }
